Py/PaddleGAN-develop/requirements.txt
tqdm
PyYAML>=5.1
scikit-image>=0.14.0
scipy>=1.1.0
opencv-python
imageio==2.9.0
imageio-ffmpeg
librosa
numba==0.53.1
easydict
munch
natsort
paddlepaddle==2.0.0-rc1
dlib